+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/EnchantmentsFeature.java
@@ -97,7 +97,7 @@ public class EnchantmentsFeature extends JsonFeature {
             Changes to damaging enchantments:
             Enchantments deal bonus damage based off the item's attack damage. So Sharpness on a Sword adds less damage than Sharpness on an Axe.
             Sharpness deals +0.75 damage per level
-            Smite deals +1.5 damage per level to undead
+            Smite deals +1.25 damage per level to undead
             Bane of Arthropods has been replaced with Bane of SSSSS that deals +1.25 damage per level to arthropods and creepers and applies slowness""")
 	public static Boolean replaceDamagingEnchantments = true;
 	@Config
@@ -118,9 +118,7 @@ public class EnchantmentsFeature extends JsonFeature {
 			IdTagMatcher.newId("minecraft:feather_falling"),
 			IdTagMatcher.newId("minecraft:looting"),
 			IdTagMatcher.newId("minecraft:fortune"),
-			IdTagMatcher.newId("minecraft:luck_of_the_sea"),
-			IdTagMatcher.newId("allurement:reforming"),
-			IdTagMatcher.newId("allurement:alleviating")
+			IdTagMatcher.newId("minecraft:luck_of_the_sea")
 
 	));
 	public static final ArrayList<IdTagMatcher> disabledEnchantments = new ArrayList<>();